Rename `common` to `aliases`
