+++ web/client-js/src/store/board/mutations/board-created-mutation.ts
@@ -26,7 +26,7 @@
 
 import { Mutation } from "vuex";
 import { BoardState } from "@/store/board/state/board-state";
-import { BoardCreated } from "@/store/board/common";
+import { BoardCreated } from "@/store/board/aliases";
 
 /**
  * Mutates the local {@linkplain BoardState board state} in response
