+++ web/client-js/src/store/board/mutations/column-added-mutation.ts
@@ -26,7 +26,7 @@
 
 import { Mutation } from "vuex";
 import { BoardState } from "@/store/board/state/board-state";
-import { ColumnAdded } from "@/store/board/common";
+import { ColumnAdded } from "@/store/board/aliases";
 
 /**
  * Mutates the local {@linkplain BoardState board state} in response
